backend/requirements.txt
google-adk>=1.19.0
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
opentelemetry-api
opentelemetry-sdk
python-pptx
playwright
termcolor
python-multipart
httpx
